+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- support for `XPath` in `xml-prune` scriptlet
+  [#325](https://github.com/AdguardTeam/Scriptlets/issues/325)
 - conversion of UBO's redirect priority to the converter.
 
 ### Fixed